+++ test/bigtest/network/config.js
@@ -40,7 +40,7 @@ export default function configure() {
   this.get('/bl-users/_self', {});
   this.post('/bl-users/password-reset/validate', () => {
     return new Response(204, {}, '');
-  }, 204);
+  });
   this.post('/bl-users/password-reset/reset', {}, 401);